Remove "extra_pop" from call handling code
The CALL_METHOD opcode has to handle two cases: The one where `LOAD_METHOD` successfully "unpacked" a bound method (conceptually, in reality it happens through how we interpret the caches) and the case where we did not end up having anything to unpack.

In the later case we have an unnecessary `_Unbound` on the stack that needs to be cleaned up. Previously we used the same strategy as cpython where the arguments are removed from the caller stack before the target function is called. In this case we would just remove one extra element (hence the `extra_pop` parameters to specify how many extra elements to remove).

This style of handling things however only works when there is one stack pointer for each frame. And I want to move to only a single stack pointer for the whole thread. This requires that we deal with this `CALL_METHOD` in a different way. This adds a `removeValueAt` function to remove values in the middle of a stack frame.

While this is less efficient than the previous method, it shouldn't be too bad because it is the uncommon case of `CALL_METHOD` and it enables changes in the rest of the stack that will more than make up for it.

Based on Facebook D23636383