Clean arguments from caller stack after the call returns
To perform a call code has to push the function and all arguments to the stack and then execute a call opcode (or `Interpreter::callXXX`). Conceptually the function and the arguments now no longer belong to the callers frame but to the callees frame. The call procedure must ensure that the arguments are remove from the stack after the call returned.

Currently this happens before the call even started: Because each frame has its own stack pointer, the code can simply change the stack pointer of the caller to no longer include the function + arguments before the callee even starts running.

However having a stack pointer for each frame is wasteful and not necessary in our design (contrary to cpython which has frames spread accross the heap and thus needs a stackpointer per frame).

This changes the code to remove the arguments from the stack when `RETURN`ing from a call. This ended up being a lot of churn because all the various error conditions in a call must also cleanup all the arguments now (becaus we never reach a RETURN in that case).

I also kept the `post_call_sp` calculation around in debug builds, so we can verify that the more complicated removing on return is correct.

Based on Facebook D23636436