backport bpo-39406: Implement os.putenv() with setenv() if available (GH-18128)
This re-applies D19648559 which was accidentally lost in the 3.7 upgrade.

This backports to bpo-39406 to pyro. This fixes `os.environ` changes not reflecting in `os.execv` Original summary:

If setenv() C function is available, os.putenv() is now implemented
with setenv() instead of putenv(), so Python doesn't have to handle
the environment variable memory.

Based on Facebook D23804523