Remove __module__ attributes from non-heap types in builtins module
Extension types in cpython only have a `__module__` set when `tp_name` contains a ".". Most builtin types in cpython do not use a dot and have no `__module__` set. This emulates this in pyro by remove the `__module__` attribute from non-heap types when the value is "builtins".

This fixes the behavior of the `inspect` module when `inspect.getmodule` is applied to frame objects.

Based on Facebook D23641597