Reorder code in typeInit
Reorder code in `typeInit()` to better match the equivalent cpython code; That doesn't change any behavior but makes it easier to compare the two implementations.

This also changes the code to set `__qualname__` in the resulting type and not in the supplied dictionary (which should not be modified).

Based on Facebook D23694777