Implement built-in stubs for traceback
Add all relevant methods and attributes to the `traceback` type.

Add the slots needed in `RawTraceback` for the full implementation,
but leave the temporary string hack in the `frame` slot.

Keep `lineno` and `frame` hidden, since those will be calculated
only as needed using the `function` and `lasti` slots.

Make traceback non-subclassable.

Based on Facebook D23608186