Enable tracebacks by default
This diff turns on built-in tracebacks, and implements a writer to print
the traceback to a file, one line at a time.

Take two passes per line to make the strings for printing:
one pass to calculate the length and another to build it.

Remove the temporary `_traceback_str` hack.

Remove our internal fork of `unittest/result.py`.

Leave `sys.tracebacklimit` and repeated lines for future diffs.

Based on Facebook D23845180