Merge _type_new/_type_init and typeNew/typeInit back into one function
Summary: This merges the sequence of `_type_new`, compute mro, `_type_init` that we currently use within `type.__new__` back into a single `_type_new` call. This sets the stage for upcoming changes: We need to recompute the metaclass before the actual type object is created; currently the MRO is computed too early and needs to be moved later in the process. Builtin/bootstrap types need a way to skip mro computation, ...

Based on Facebook D23817548