Compile bytecode upfront in benchmark_rev --run-benchmarks mode
Summary: The `benchmark_rev` / `diffrevs` mode to run the microbenchmark would start compiling the bytecode, skewing the results with the work necessary to compile the code. In order to focus on the actual benchmark and to match the behavior of the pyro-perf CI job compile the bytecode upfront.

Based on Facebook D23963717