Remove credits, help, license, quit from builtins.py
`credits`, `help`, `license`, `quit`, `exit` and `copyright` are added to the `builtins` module by `site.py`, they are not naturally present. We should not add them in `builtins.py` to match cpython behavior.

They are not missing when the site module isn't run which matches cpython.

Based on Facebook D23963521