typeLookupInMro tweaking
Nowadays `typeLookupInMro` is guaranteed to only have interned string keys, which guarantees us that there are no python function calls for comparison and hashing and hence there cannot be any garbage collection cycles. This mostly with `typeLookupInMroById` as we can skip the `HandleScope` there.

This function is performance relevant so we can do the rare step of changing the interface to `RawXXX` parameters.

Based on Facebook D23733309