Remove addEmptyBuiltinType
The function `addEmptyBuiltinType` sounded like a convenience function that does the same as `addBuiltinType` without attributes. However confusingly it was also setting the layout to enable tuple overflow. This made sense for all the dozens of exception classes where it was used before D23817756, but it was wrong for all the other users of the function.

With the newly added `kNoAttributes` constant there is nearly no convenience left of using the function and the danger of unexpected overflow behavior. Therefore this removes the function.

Based on Facebook D23817780