Use newInstanceWithSize in newLayout and newType
`newLayout` and `newType` manually sized and allocate a new object, however they are just normal instances so we can just as well use `newInstanceWithSize` for them to simplify the code.

There was one oddity with `RawLayout::initialize`: It would initialize the hash field in the object header with the layout id. However this isn't used anywhere right now so we can remove it.

Based on Facebook D23962750