Fix results when BINARY_OR_SMALLINT is reverted
This fixes a bug introduced in D23820052: BINARY_OR_SMALLINT getting called without smallints would correctly revert to the different opcodes, but would fail to compute the correct result in that case, since the original value was overwritten by the type checking logic.

This also overhauls the tests for `BINARY_XXX_SMALLINT`.

Based on Facebook D23990575