Skip membership check when caching global variables
When a global variable is cached in a function, the function's all `LOAD_GLOBAL/STORE_GLOBAL` get rewritten to use the cache for the same global variable,  and the function is added to the dependency list of the global variable for cache invalidation later.

The problem is that we currently maintains the list as set, to guarantee that all functions in a dependency list are unique by performing membership check

This invariance turns out to be costly and too strict:

- costly: imagine a module file with 1k functions where all the functions access a same global variable. The global variable's dependency list will contain all functions and linear scan of such a list will have time complexity of O(n) where n is the number of functions in the module file.

- too strict: caching a specific global variable in a function happens only once. Therefore, the dependency list mostly consists of distinct functions.

Based on this understanding, this diff drops the membership check when a function is added to a global variable's dependency list at the cost of allowing duplicate functions in the dependency list of a function. It can happen in the following scenario

```
def foo():
  import sys
  t = sys.modules[__name__].global_var # a
  ...
  return global_var # b
```

In the function `foo`, `global_var` is cached twice first via `LOAD_ATTR` at line a, and second via `LOAD_GLOBAL` at line b.

By dropping membership check, `foo` is added to `global_var` twice, ended up being duplicated in its dependency list.

Note that this duplication does not hurt the correctness of cache invalidation, but may slow it down by processing same functions multiple times.

Based on Facebook D24071784