Assume fixed nargs in set.intersection implementation
Ever since the builtin function system was introduced the `nargs` parameter of a given function is fixed to the number of parameters specified in the function definition in the python file. For `set.intersection` this is `def set(self, other):` resulting in `nargs == 2`. Remove invalid tests and code paths for `nargs != 2` that would never get executed.

Also move the `TODO(T46058798)` about actually using a varargs parameter for `intersection` to `builtins.py`.

Based on Facebook D24113426