Use object address for C-API table hash
We have found that C-API code tends to allocate handles of a lot of `str` and
`bytes` objects, often in calls to `os.stat`, which goes through the C-API
posixmodule. Since our previous hash function for the C-API was
`Runtime::hash`, which hashed the *values* of those `str` and `bytes` objects,
we often had enormous strings of collisions in the C-API handle table. This led
to such a degradation in performance ("a little under 50% of CPU time in
distillery startup spent in two functions: `py::identityDict::atPut` and
`py::identityDict::at`") that we needed to reconsider the hash function.

This diff changes the hash function to a instead be based on the pointer of
objects in the handle table. This means that equivalent objects will not
necessarily land near one another in the table. Since hashes become so cheap
to compute (cheaper than a load), don't store them in the table anymore.

Since we have a moving garbage collector, object pointers change with every
GC. This diff therefore rehashes the whole table after each GC.

Based on Facebook D24122637