[bootcamp][pyro] Add support for marshal dump version 3
Implements support for `marshal` dumps with version 3 in Pyro.

The main difference between version 2 and version 3 is that the protocol version 3 allows for back references in case the same object is referenced multiple times (TYPE_REF / FLAG_REF).

The implementation makes use of a dictionary to keep track of objects references doing an initialization pass.

Later, when dumping the objects, those which are referenced more than once are assigned an appropriate address which is later reused to reference them instead of dumping the whole object again.

This diff also adds unit tests for marshal version 3 `dump` and `dumps` which replicate the existing one for version 2 but using the expected output provided by Python 3.7.5 marshal's implementation.

Based on Facebook D24197843