Create default tp_new method that can be returned by PyType_GetSlot for builtin types
This implements a generic `tp_new` method called `superclassTpNew`, that looks up a type's `mro` to look for a `__new__` method to call. It also handles `PyObject` allocation and `NativeProxy` initialization when needed. This function can also be returned as the result of `PyType_GetSlot` on builtin types, since any later-passed in type to the function must subclass the original builtin, and therefore have it in its `mro` chain.

Since `superclassTpNew` has mostly the same functionality as `PyType_GenericAlloc` and `PyObject_Init`, this diff also splits that functionality into separate functions that can be used in both cases.

Based on Facebook D24144179