Expose builtin types Exception and Type
In order to fully support subclassing `Exception` and `Type` in Cython, we need to expose them as non-opaque types.

While this technically goes against the contraints of the Limited API specified in PEP-384, it has been deemed an acceptable temporary divergence in order to quickly get `thrift` to compile and run.

Based on Facebook D24684353