Set intrinsics via docstring, not array
Read frozen function docstrings for the magic `"$intrinsic$"` in the first
line. If functions are marked as intrinsic, a prototype is generated in
`build/bootstrap/builtins.h` and added to the list in
`build/bootstrap/builtins.cpp`.

This removes a need for a separate intrinsic array (which removes the last
piece of code referring to intrinsics by SymbolId) and enables methods to have
intrinsics.

Based on Facebook D24718362