Remove UNIMPLEMENTED from _Py_Dealloc for types without slots
`_Py_Dealloc` was changed to be `UNIMPLEMENTED` instead of a no-op in D18718315 as a signal that we don't support the function.

However, since the function was removed from CPython in https://bugs.python.org/issue39543 as a result of not being used, I think we can safely keep it as a no-op when called with a type that doesn't define a `tp_dealloc` slot.

Based on Facebook D24621419