Move C-API state out of runtime
Move the `ApiHandle` tables from the runtime to the C-API.

Use a static buffer to store the state, with size statically guaranteed to fit all of the necessary data.

This change will enable us to move all `ApiHandle` implementation details out of the runtime in a future diff.

Based on Facebook D24622060