Add a shortcut for `list(sth)` expressions
This change adds `_list_ctor` to be invoked for `list(sth)` expressions instead of following a roundabout path of `type.__call__`, to improve the performance of distillery.

This change is made according to @matthiasb's python profile data showing that `list.__new__` is the most frequently used callees from `type.__call__` from distillery as of now.

Based on Facebook D24804796