Move modules-by-index list to C-API
Convert `modules_by_index_` from a list of module objects to a vector of `PyObject*`,
and move it from the runtime into the C-API state.

This diff results in 58.4% fewer calls to `ApiHandle::borrowedReference` in `test_wsgi` (from 4.11M to 1.71M). This caused a 0.46% overall decrease in instructions.

Based on Facebook D24601296