[codemod] apply pyfmt with usort to opted-in sources
Facebook:

µsort is now enabled by default in pyfmt:
https://fb.workplace.com/groups/pyfmt/permalink/699617850675734/

This applies usort to all remaining files opted-in to pyfmt.
As µsort is safer than isort, this should not cause major issues,
but as µsort ignores isort:skip_file, this will affect files
that have previously been unsorted.

If this breaks your project, please do not revert this diff.
Please fix forward and/or apply usort:skip on lines that must stay
in place. If you have modules with import-time side effects, please
talk to Python Foundation and we will add your modules to the global
configuration, so that you don't need to use skip directives.

µsort user guide:
https://usort.readthedocs.io/en/stable/guide.html

FAQ about µsort in pyfmt:
https://www.internalfb.com/intern/wiki/Python/Code_Formatting/%C2%B5sort/

@drop-conflicts
@allow-large-files

pokemon_party

Based on Facebook D24880203