Rework thread interrupt / pseudo stack overflow handling
We already use pseudo stack overflows to interrupt threads to turn unix signals into exceptions. So far we always aborted the frame setup when this happened. This extends the system with the option to continue frame setup after performing the interruption work.

- For every `Thread::pushXXX` add a `Thread::handleInterruptXXX` function, so we can keep all the slow-paths in a separate code path. (Without this I would get suboptimal code generation as the C++ compilers appeared to place some spill-reload instructions into the fast-path).
- Add `Thread::interrupt_flags_` word to thread, so we can set/clear possible reasons for a thread interruption independently. This is currently `kSignal` for unix signals; `kProfiling` for profiling will be added soon.
- Add `Thread::handleInterruptWithFrame` hook to perform extra operations after the new stack frame was setup. This will be used by the profiler.

All of this does not affect performance in any measurable way.

Based on Facebook D24548797