Encode beginning of locals with an offset instead of a pointer
Change callframes to encode the beginning of the locals by an offset relative to the frame instead of a pointer. This turns out to be as-fast when accessing locals (most locals are accessed with `reverseLocal()` anyway and don't require the pointer, the remaining cases could all absorb the addition in an x86 addressing mode). This saves a couple instructions when setting up frames and saves translating the pointer when moving generator/coroutine frames from the heap and back.

- Remove `Frame::init()` function. We had a split where parts of the frame was initialized by `Frame::init()` and another part in `Thread::pushXXX`. Removing `init()` simplifies the code as all `Frame::setXX()` calls are now in the `thread.cpp`.
- This changes `Frame::locals()` to point to the actual point where locals begin rather than 1 word short of it as we did previously. That means local 0 is now found at `locals()[-1]` instead of `locals()[0]`, local 1 at `locals()[-2]`, etc. The end address is easier to compute when setting up the frame and there are practically no locals() accesses left without a constant, so things get folded anyway.
- Add `Frame::setLocalsOffset()`/`Frame::localsOffset()`.
- We previously needed to "stash"/"unstash" the stack size and locals pointer when moving frames from and to the heap. This is only required for the stack size now and this change accordingly renames `stashInternalPointers` to `stashStackSize`.

Based on Facebook D24505894