Introduce return mode to call frames
This introduces the concept of a "return mode" that is stored with every call frame. This allows to modify the behavior when a function returns.

This first version introduces the `kNormal` return mode and `kExitRecursiveInterpreter` return mode which is used instead of the comparison with an `entry_frame` to determine whether to exit a recursive interpreter call.

I expect to use the "return mode" for things like:
- Notify a profiler that execution is leaving a function.
- Do things like popping a value from the stack or doing a type check to allow more situations in the interpreter to use tailcalls.

Checking the return mode also saves a couple instructions compared to the previous approach of comparing with `entry_frame`.

Based on Facebook D24511500