Fix crash for non-utf8 python files
Do not use `Runtime::newStrWithAll` on untrusted user input that may not be UTF-8. This fixes crashes when running with non-UTF8 files.

Note that there are odd behavior differences in cpython: Where parsing from a `bytes` object does not check for valid UTF-8 outside of string literals, while parsing a `FILE*` does check the whole file for UTF-8 validity. You will notice that you can `import` such files in cpython but not use them as `__main__`. I think as long as we don't crash with can live with the differences given that cpython itself isn't super consistent with the `import` vs. commandline differences.

Based on Facebook D24902873