Increase wait time for thread tests to decrease flakiness
I think I saw this test fail a bunch lately. I suspect the time may not be enough for the threads to finish running in all cases. Let's increase it.

(Eventually we should have some form of thread sync so we can fix this properly without relying on time)

Based on Facebook D24903042