Remove RawLargeStr::kDataOffset
Summary: It seems someone wanted to prepare `RawLargeStr` to potentially have extra info before the string data starts. However the field currently is 0 and only a single user used this field and all the other users accross the whole codebase ignored it. So I'd rather remove this.

Based on Facebook D25003372