Switch opcode handler tables to enable/disable opcode counting
This introduces mechanisms for switching the handler table of the assembly interpreter at runtime. I needed a way to enable/disable opcode counting for the profiling code and took this as an opportunity to explore what it would take to switch out the handler table of the assembler interpreter at runtime[1].

Emitting multiple versions of the handler table in the assembly interpreter is not hard. The trick here is to update the `kHandlerBase` register to get the updates in place.

I realized that for a foolproof solution that works in any circumstance we would need to "restore" the value of the `kHandlerBase` register on every `emitRestoreInterpreterState`. My measurements indicated that this would cost around 0.5% cg_instructions. Given that this is not a critical feature yet I went for a more lightweight solution that only updates `kHandlerBase` on thread interruptions ("pseudo stack overflow"s) reducing the cost to 0.1% cg_instructions. This unfortunately means the effects won't be felt when there are recursive interpreter invocations on the C++ stack at the time the handle table is switched. The loops further up the stack will likely miss the update. I decided that this is acceptable for the profiler for now, in combination with a warning printed when the profiler is enabled in a deep callstack with non-recursive invocations so the user knows when to expect inaccuracies.

[1] While this is a neat trick for the profiler; the real use for this would be implementing an optional read barrier that can be switched on/off by the GC. (That will be some way down the road, assuming we go with read-barriers and parallel GC at all).

Based on Facebook D24762899