Change Thread::exec to set missing __builtins__ to a module proxy
Change the code in `Thread::exec` that adds a `__builtins__` entry to globals if it is missing to use the module proxy of the builtins module (rather than the module itself) to match cpython behavior.

We got confused when implementing this earlier because cpython (accidentally?) puts in the module in the case of `__main__` but nowhere else.

Based on Facebook D24998890