[PEP-384 _datetime] Remove limited API calls
Remove the remaining two API calls that we cannot support in Pyro.

Migrate `_PyObject_GetDictPtr` to `PyObject_GenericGetDict`,
which is already a thin wrapper around `_PyObject_GetDictPtr` in CPython.

Migrate `_PyUnicode_Copy` to `PyUnicode_AsUCS4Copy`,
which requires one additional allocation and copy,
but eliminates the need to mutate a string.
Since this is only used to replace surrogate code points
with a `'T'`, this should not adversely affect performance.

Based on Facebook D25101084