Patch type's layout when setting type's __class__
Our original implementation of setting a type's `__class__` would set it's own layout to be the new class's layout. However, this makes any old instances of that type un-inspectable and can cause memory issues.

The proper behavior should be creating a new layout that looks exactly the same and setting the type's layout to the newly created layout. This diff does so by making use of the function `layoutCreateChild`. This diff also makes that function public and renames it to `layoutCreateCopy` to make it more immediately obvious what the function does.

Based on Facebook D25230193