Copy builtin function pointer to function object
Rename `Function::stacksize` to `Function::stacksizeOrBuiltin` and in case of builtin functions copy the function pointer to the builtin into this field, so we don't need an indirection to the code object when calling.

Also emit an `ALIGN_16` specification for builtin functions and use `SmallInt::fromAlignedCPtr` to save a shift instruction when using the field.

Based on Facebook D25214809