Create PyObject without native data via PyType_GenericNew
This change fixes a bug that `PyType_GenericNew` mistakenly assumes that any `type` given to `PyType_GenericNew` satisfies `type.hasNativeData` which is untrue when the `type` object's size is less than a certain threshold by D25093569. D25093569 made the same change to `PyObject_New`, but not to `PyType_GenericNew`.

Note this this fix was inspired by a crash observed from Distillery: `datetime.tzinfo` is created with the slots:

```
static PyType_Slot tzinfo_slots[] = {
    {Py_tp_doc, tzinfo_doc},
    {Py_tp_getattro, PyObject_GenericGetAttr},
    {Py_tp_methods, tzinfo_methods},
    {Py_tp_new, PyType_GenericNew},
    {0, 0},
};
```

and is subclassed by a managed type in `pytz` module, causing a crash.

Based on Facebook D25379649