Avoid masking in type check for Error, Bool and singletons
We usually have to mask out the `kImmediateTagMask` lowest bits to check whether an immediate is a specific type.

However in the case of `RawNoneType`, `RawUnbound` and `RawNotImplemented` are all bits known upfront and we could for example check the whole lowest byte instead to skip the masking operation. Moreover in the case of `RawError` and `RawBool` we can move the true/false and error kind information to the 2nd byte and also check the whole lowest byte without masking.

This changes `Bool::trueObject` and `Bool::falseObject` to use more than 8 significant bits so they no longer fit the `LOAD_IMMEDIATE` opcode. This leads to a small regression in the richards benchmark. This will be recovered in the next change that adds a `LOAD_BOOL` opcode.

Based on Facebook D25331146