Move codePointAt, codePointLength, offsetByCodePoints from RawLargeStr to RawDataArray
Move `codePointAt`, `codePointLength` and `offsetByCodePoints` from `RawLargeStr` to `RawDataArray`. They usefull for byte buffers as well in situations in which the buffer is known to contain UTF-8 or where wrong answers in of invalid UTF-8 is acceptable.

This also splits the `codePointAt` implementation into specialized `RawSmallStr` and `RawDataArray` variants which should help performance.

Based on Facebook D25212933