pyrobench: Keep callgrind output files around in --run-benchmarks
When analyzing the performance differences between two runs it is often helpful to compare callgrind output before and after. This changes `diffrevs` to keep the callgrind files around in `/var/tmp/benchresults` instead of just using temporary files. That makes it easier to grab the traces after the runs and analyze them in detail.

- Add option to pyro/benchmarks/run.py to specify a directory to place callgrind output files and leave them there when the script returns.
- Use the new option in pyrobench/benchmark_rev.py (which is used by diffrevs.py)

Based on Facebook D25401921