Avoid __new__ lookup from type.__call__
The currently implementation of `type.__call__` always performs lookup of `__new__` in the given `cls` object. The lookup result is equal to `object.__new__` most of the time for which `type.__call__` creates an instance of the given type natively.

This change avoids such type lookups. `type` objects are populated with `Type::Flag::kHasObjectDunderNew` flag: if set,
it entails that the type object's MRO lookup for `__new__` is guaranteed to be equal to `object.__new__`.

Based on Facebook D25725020