[codemod][fbcode][2/n] Apply clang-format update fixes
Facebook:

D25840887 updates `arc lint`s clang-format to 11.0.1. There are some
minor formatting changes that come with that and this fixes them.

It is important that these changes be made to avoid diff-time lint noise
as a result of the update, but more importantly, so that the continous
formatting that happens daily via CodemodService (e.g D25845127) does
not attempt to fix all of these issues at once.

This diff is a subset of the changes produced by
`tools/arcanist/lint/codemods/clangformat-fbsource`.

#nocancel
@drop-conflicts
@override-unit-failures
@allow-large-files

Based on Facebook D25861960