Add shortcut for __import__ when the module exists in sys.modules
Summary: Add a new builtin to _imp.py which checks if the module exists in Runtime::modules and returns it if present. If it is not present it returns None to allow the regular import mechanism to load the module. This shortcut avoids locking when re-importing a module that has already been loaded.

Based on Facebook D25097684