Return True for bool(obj) where type(obj).__bool__ and type(obj).__len__ are nonexistant
Our runtime performs type lookups for `__bool__` & `__len__` when

```
if obj:
  ...
```

gets executed when `obj` is not a builtin type. This is very expensive and this cost gets exacerbated due to the fact that most `obj`'s types do not have `__bool__` and `__len__` at which the runtime just returns `True`.

This change marks user-created types with custom `__bool__` and `__len__` so that `bool(obj)` can be immediately evaluated to `True` when the type of `obj` does not have any such dunder functions.

Based on Facebook D25958512