Return False from isinstance_intrinsic(obj, type) when type(obj).__class__ is object.__class__
Currently the intrinsic code of `isinstance_intrinsic(obj, type)` returns `False` immediately when the type of the `obj` is not a subtype of `type` and the type of the `obj` is a builtin one.

This change extends the idea to cover all types whose `__class__` just returns the type of the given object. This includes

- All types that do not redefine `__class__`

`NoneType` is a special case since Pyro adds `NoneType.__class__` as a hack, but its behavior is the same as `object.__class__`. This change includes this special case to enable `isinstance(None, type)` to return False in the intrinsic.

Based on Facebook D26021166