Remove Bucket class from IdentityDict
We are moving away from "buckets" of data in a single array
to a sparse/dense split with arrays for indices, keys, and values.

This diff moves a number of helper functions to static scope,
eliminating the need for a blank namespace.

Also shift to using a `struct IndexProbe` for looping over indices,
and use a `uword` to reflect the address or raw value used for hashing.

Based on Facebook D25875089