Handle slot_descriptor.__get__ owner parameter
The `__get__` method on a descriptor accepts a second argument, which is the owner class. See:
- https://docs.python.org/3/howto/descriptor.html#descriptor-protocol
- https://docs.python.org/3/reference/datamodel.html#object.__get__

This second argument is optional and defaults to `None`, but if the first argument to `__get__` is `None`, then the second argument should be required. But currently, nothing is enforcing that the second argument has been provided.

This change passes the `owner` parameter through to the `py::slotDescriptorGet` function, and if both arguments are `None`, the appropriate `TypeError` is raised with message `"__get__(None, None) is invalid"`. If the first argument is `None` and the second argument is valid, then the descriptor is returned and no error is raised.

Corresponding tests are added to `builtins_test.py`.

Because this change also touches the `__delete__` method, tests are also added for this, and a related implementation issue is fixed to ensure that the proper error is raised when `__delete__(None)` is called.

Based on Facebook D26262982