Add ScratchReg helper
Introduce `ScratchReg` helper that automatically grabs a register that is unassigned in the register state tracker and frees it in the destructor when leaving the scope.

Use it across `interpreter-gen-x64.cpp`. I think none of the instructions uses a direct register reference anymore now; it's all `ScratchReg` or `VirtualReg`.

Based on Facebook D25214873