Introduce Byteslike handle
This introduces a new `Byteslike` handle that allows uniform access to byteslike objects. Byteslike objects are objects that implement the buffer protocol / `PyBufferProcs` in CPython.

The trick here is to have the layout of the class match the layout of a `Handle<Object>` so we can optionally register it with a `HandleScope` in case the underlying object is based on a `LargeBytes` or similar managed object that may be moved around. Other objects like pointers into the unmanaged heap or into SmallBytes are taken as-is and not registered in the `HandleScope` but the address is offset by `kHeapObjectTag` so we can uniformly all variants.

The protocol to use the handle differs slightly from other handles in that you have to explicitely need to call an `init()` function and get a `true`/`false` result back depending on whether the passed object is a `byteslike` object or not. This avoids inefficient code where we would first have `Runtime::isByteslike()` check for a number of known byteslike cases while the handle constructor needs to perform all those checks anyway and we don't want to perform them twice.

As a proof of concept, this changes the implementation of `bytes.__iadd__` to use the new handle.

Based on Facebook D26383349