Store results of file-accessing functions in the correct data types
The `::read` and `::lseek` functions return a `ssize_t` and `off_t` respectively, which are defined as 64-bit types on a 64-bit machine. Storing them in the 32-bit `int` type has led to signed-integer overflow on the large files created by `buck`. This diff ensures that the calls to these `stdlib` functions match the signature specified in their `man` pages.

I've also vetted the rest of the `File` functions to ensure that all of them store the results of their `stdlib` calls in the correct data type.

Based on Facebook D26484212