Add support for bytes-backed memoryviews to `PyObject_GetBuffer` API
Adds support for returning the underlying buffer of byte-backed memoryviews for the PyObject_GetBuffer API. Matches CPython for refcount behavior. Since all byteslike objects get normalized into this form when we construct a memory view, this supports bytearrays, etc. by default, as demonstrated in the tests.

Not implemented: Support for raw pointer backed memory views. Will be done in a future change.

Based on Facebook D26486745