Manually strip whitespace when calling `int.__new__` with bytes
Context: In the future, we plan on extending the `int.__new__` support to all byteslike objects. Some of these byteslike objects such as `array.array` have no support for `.strip()`.

Towards this end, this diff changes `intFromBytes` to rstrip the bytes object given to it as well, and removes the `.strip()` calls in the builtins module.

Based on Facebook D26515247