Support str.format({self}, self="name")
Summary: Turns out that the CPython string formatting uses a clever trick of not explicitly passing self in order to support `self` getting passed in as a keyword. This was causing divergence between Pyro and CPython's behavior, so let's use the same trick and preserve error messages in str.format().

Based on Facebook D26553790