Support dict.__reversed__ and dict_keys.__reversed__
Summary: This diff adds support for the CPython 3.8 features of `__reversed__` for dicts. The approach is deliberately naive: The goal is to support the behavior without investing too much energy into it immediately - if this ends up being expensive for our distillery use case in practice, we can go the route of CPython and create a special `dict_reversekeyiterator` native class and simply modify the implementation of `dict_keys.__reversed__`, but let's not implement the special iterator class until there's a proven need for Instagram.

Based on Facebook D26585860