Handle buffer protocol objects in memoryview.__new__
Add `newBytesFromBufferProtocol` function in the C-API and export it
to the runtime. This is a bit of a hack and **should not be used** long-term.
It creates a copy of data from a buffer protocol object and stores it in a
`bytes`, with intent to store in a `memoryview`.

In the future we should instead be able to directly `bufferinfo` objects from
memoryviews.

Based on Facebook D26593092