Add flags to optionally suppress abort on _unimplemented
By default, we want _unimplemented() to abort because we want to "Fail Fast"

During development and testing, it can make more sense to generate a Python exception so that a program can keep running.

Your mileage will vary, the code is developed with the assumption that execution will not proceed past a call to _unimplemented!

Based on Facebook D26754099