Avoid recursive tp_new calls with extension objects
Summary: When looking for Py_tp_new method superclassTpNew skip parents with slots implemented. We are able to do this because on type creation if a tp_new slot existed on a parent class it would have been inherited. Since we are in superclassTpNew we can safely assume that there were no slots to inherit.

Based on Facebook D26765902