[rfc] Introduce script to build the `pyro.instagram.server` FBPkg
Summary: Context: We're looking to get Pyro working with the Instagram server orchestrator integration tests. The way the server orchestrator integration tests work is by fbpkg.fetch'ing `instagram.server:DEFAULT`, which is a squashed version of distillery. In the long term, we should be able to just use that, but since we're not there yet, let's create a FBPkg  that squashes pyroinsta in the same manner so that that work is unblocked on removing the distillery symlink + monkeypatches from Pyro for now.

Based on Facebook D26896498