Consistently call resetPythonEnv before Py_Initialize in tests
Make sure `resetPythonEnv()` is always called before `Py_Initialize()` in C-API tests. This usually happens automatically in the test fixture but a couple tests with custom fixture did not do so.

CPython has a bunch of global state and depending on the test order that global state may be off. The `resetPythonEnv()` function should reset all the things that may cause trouble.

Based on Facebook D26955870