Add back par_style = xar to benchmark builds
Summary: The benchmark server currently complains about not being able to run `_compile_tool`. This is probably because the `par_style = "xar"` line was removed for `benchmark.server` in D26929344. It think we need it for the client / server builds which package tools that are run externally, but should be fine to not have it for `upload_to_phab` and `format_results`.

Based on Facebook D26983654