Add a CMake mode to copy frozen modules instead of re-generating them
Problem: Pyro doesn't build in TP2 right now. The basic reason is that util/freeze_modules.py depends on Python 3.8, whereas Python 3.8 isn't available in TP2 (and, in general, it is a bit tricky to depend on the Python version in there, and can cause reproducibility issues).

We discussed how to best address this issue with Matthias and Max, and there were a couple possible options:

1) always check in generated files for FBCode + TP2
2) wait for the TP2 python version to be what we need, which can cause problems in future version upgrades as well
3) compromise, and check in the generated frozen module files to TP2, while keeping our current one-step build system for regular development

The consensus was that we should go with (3), as checking those files in breaks the python dependency in TP2 while avoiding a bad DevX. This diff lays the groundwork for that by adding a flag to CMake that allows us to copy files from a builtin location.

Next up: We will modify `util/update_tp2_sources.py` to also build & copy the frozen module files, and change rules.py in tp2 to pass in the COPY_FROZEN_MODULES option.

Based on Facebook D26984462